Moved content management to content controller and content services. Changed routers. Added ArgumentError exception to ApplicationController, it triggers when an invalid value is entered in the Content model position. I left authorized checking only in update, destroy and content management methods.

+++ app/services/screens/delete_screen_service.rb
@@ -1,17 +1,20 @@
 # frozen_string_literal: true
 
 module Screens
-  class DeleteScreenService < ApplicationService
+  class DeleteScreenService < AuthorizedService
+    attr_reader :id
+
+    def initialize(current_user:, id:)
+      super(current_user:)
+      @id = id
+    end
+
     def call
-      screen = Screen.find_by(id:)
+      screen = current_user.screens.find_by(id:)
 
       return OpenStruct.new(success?: false, screen: nil, errors: ["screen not found"]) if screen.blank?
 
-      begin
-        authorized screen, :destroy?
-      rescue StandardError
-        return OpenStruct.new(success?: false, screen: nil, errors: ["screen not found"]) if screen.blank?
-      end
+      authorize screen, :destroy?
 
       screen.destroy
